All set up and working.
What we have here is paletton.org application code pretty printified.

As research shows, the code is basically a bunch of modules minified together using requirejs.
This makes it rather difficult to understand. I created this small package in order to make it
easier to research the internals of said application.

1. app.compiled.js is what you get when you visit paletton.com.
2. app.js is the same passed through a pretty printer -- unminified, that is.
3. Run code.coffee to obtain a 'modules' directory where you may read each module definition
separately.

Enjoy!